+++ Cargo.toml
@@ -255,6 +255,7 @@ codesort = "1.0.0"
 lazy-regex = "3.4.1"
 termimad = "0.33.0"
 fdt = "0.1.5"
+polonius-the-crab = "0.4.2"
 
 # data structures
 bilge = "0.2.0"
@@ -274,6 +275,9 @@ enum-map = "2.7.3"
 regex = "1.11.1"
 h2 = "0.4.6"
 zstd = "0.13.2"
+# used in pcode cache
+# we don't need ahash, sync (parking_lot), or stats so disable here
+quick_cache = { version = "0.6.14", default-features = false }
 
 # macros
 proc-macro2 = "1.0.89"
