+++ styx/core/styx-cpu-pcode-backend/Cargo.toml
@@ -57,6 +57,7 @@ tap = { workspace = true }
 styx-workspace-hack = { version = "0.1", path = "../../workspace-hack" }
 derive_more = { workspace = true }
 smallvec = { workspace = true }
+polonius-the-crab = { workspace = true }
 
 [dev-dependencies]
 keystone-engine = { workspace = true }
