+++ styx/core/styx-cpu-pcode-backend/styx-pcode-translator/src/lib.rs
@@ -1,17 +1,448 @@
 // SPDX-License-Identifier: BSD-2-Clause
 //! Provides a pcode translator.
 //!
-//! The three main components of the translator are the [PcodeTranslator], the
-//! [sla], and the [Loader].
+//! The three main components of the translator are the [PcodeTranslator], the [sla], and the
+//! [Loader].
 //!
-//! The [Loader] defines how the translator gets bytes to translate. The [sla]
-//! provides a specification for the processor. With both of these you can
-//! create a [PcodeTranslator] and create pcodes with
-//! [PcodeTranslator::get_pcode()].
-mod translator;
+//! The [Loader] defines how the translator gets bytes to translate. The [sla] provides a
+//! specification for the processor. With both of these you can create a [PcodeTranslator] and
+//! create pcodes with [PcodeTranslator::get_pcode()].
+//!
+//!
+//! ## Caching
 
-pub use styx_pcode_sleigh_backend::{Loader, LoaderRequires, SleighTranslateError, VectorLoader};
-pub use translator::{ContextOption, PcodeTranslator, PcodeTranslatorError};
+pub use styx_pcode_sleigh_backend::{Loader, SleighTranslateError, VectorLoader};
+
 pub mod sla {
     pub use styx_sla::*;
+}
+
+mod cache;
+
+use rustc_hash::FxHashMap;
+use std::collections::HashMap;
+use std::hash::Hash;
+use styx_cpu_type::{
+    arch::{
+        backends::{ArchRegister, ArchVariant},
+        ArchitectureDef,
+    },
+    ArchEndian,
+};
+use styx_pcode::pcode::{Pcode, SpaceInfo, SpaceName, VarnodeData};
+use styx_pcode::sla::SlaSpec;
+use styx_pcode_sleigh_backend::{NewSleighError, Sleigh, UserOpInfo};
+use styx_sla::SlaRegisters;
+use tap::TapOptional;
+use thiserror::Error;
+use tracing::{debug, trace};
+
+use crate::cache::MaybePcodeCache;
+
+#[derive(Error, Debug)]
+#[error(transparent)]
+pub struct PcodeTranslatorError(NewSleighError);
+impl From<NewSleighError> for PcodeTranslatorError {
+    fn from(value: NewSleighError) -> Self {
+        Self(value)
+    }
+}
+
+/// Easy to use pcode translator.
+///
+/// Takes ownership of a [Loader] to get bytes to translate to pcode.
+///
+/// Sets default context variables based on architecture.
+pub struct PcodeTranslator<L> {
+    sleigh: Sleigh<L>,
+    registers: HashMap<ArchRegister, VarnodeData>,
+    registers_rev: FxHashMap<VarnodeData, Vec<ArchRegister>>,
+    /// Holds a map between register varnode offsets and register names.
+    register_offset_map: HashMap<u64, String>,
+
+    user_ops: Vec<UserOpInfo>,
+
+    cache: MaybePcodeCache,
+}
+
+impl<L: Loader + 'static> PcodeTranslator<L> {
+    /// Creates a new PcodeTranslator.
+    ///
+    /// The `arch` should match the given `SlaSpec`.
+    ///
+    /// An error value indicates an error in the sla spec or sleigh backend and
+    /// cannot be handled gracefully.
+    pub fn new<S: SlaSpec + SlaRegisters>(
+        arch: &ArchVariant,
+        loader: L,
+    ) -> Result<Self, PcodeTranslatorError> {
+        let sla = S::spec();
+        let spec_file = styx_util::bytes_to_tmp_file(sla);
+
+        let arch_def: Box<dyn ArchitectureDef> = arch.clone().into();
+        // must be a reference to the NamedTempFile, otherwise it gets dropped before initializing
+        let mut sleigh = Sleigh::new(loader, &spec_file)?;
+
+        let registers = Self::get_registers::<S>(arch_def.as_ref(), &mut sleigh);
+        let registers_rev = reverse_hash(&registers).into_iter().collect();
+        let register_offset_map = sleigh.get_register_offset_map();
+
+        let user_ops = sleigh.get_user_ops();
+
+        let cache = MaybePcodeCache::should_cache(true);
+
+        Ok(Self {
+            sleigh,
+            registers,
+            registers_rev,
+            register_offset_map,
+            user_ops,
+            cache,
+        })
+    }
+}
+
+fn reverse_hash<K: Clone, V: Clone + Eq + Hash>(map: &HashMap<K, V>) -> HashMap<V, Vec<K>> {
+    let mut new: HashMap<V, Vec<K>> = HashMap::new();
+    for (k, v) in map.iter() {
+        new.entry(v.clone()).or_default().push(k.clone());
+    }
+    new
+}
+
+impl<L: Loader + 'static> PcodeTranslator<L> {
+    /// Translates an instruction at an address into a list of resulting pcodes and the number of
+    /// bytes consumed. Appends to the give pcodes vector.
+    pub fn get_pcode(
+        &mut self,
+        address: u64,
+        pcodes: &mut Vec<Pcode>,
+        data: &mut L::LoadRequires<'_>,
+    ) -> Result<u64, SleighTranslateError> {
+        let translated = self.get_pcode2(address, data)?;
+        for pcode in translated.pcodes {
+            pcodes.push(pcode.clone());
+        }
+        Ok(translated.bytes_consumed)
+    }
+
+    /// Translates an instruction at an address into a list of resulting pcodes and the number of
+    /// bytes consumed. Appends to the give pcodes vector.
+    pub fn get_pcode2(
+        &mut self,
+        address: u64,
+        data: &mut L::LoadRequires<'_>,
+    ) -> Result<TranslatedPcode<'_>, SleighTranslateError> {
+        self.cache.translate(&mut self.sleigh, address, data)
+    }
+
+    fn get_registers_option<'a, Sla: SlaRegisters>(
+        arch_def: &dyn ArchitectureDef,
+        sleigh: &'a mut Sleigh<L>,
+    ) -> impl Iterator<Item = (ArchRegister, Option<VarnodeData>)> + 'a {
+        let strings = arch_def
+            .registers()
+            .registers()
+            .into_iter()
+            .map(|r| (r.variant(), Sla::translate_register(&r)));
+
+        strings.map(move |(register, name)| (register, sleigh.get_register(&name)))
+    }
+
+    fn get_registers<Sla: SlaRegisters>(
+        arch_def: &dyn ArchitectureDef,
+        sleigh: &mut Sleigh<L>,
+    ) -> HashMap<ArchRegister, VarnodeData> {
+        let register_map = PcodeTranslator::get_registers_option::<Sla>(arch_def, sleigh)
+            .filter_map(|(register, varnode_opt)| {
+                varnode_opt.map(|varnode| (register, varnode)).tap_none(|| {
+                    debug!("Register not found in pcode: {register}");
+                })
+            })
+            .collect();
+
+        trace!("Pcode registers: {:?}", &register_map);
+
+        register_map
+    }
+    /// Gets the [VarnodeData] of register if it exists.
+    pub fn get_register(&self, register: &ArchRegister) -> Option<&VarnodeData> {
+        self.registers.get(register)
+    }
+
+    pub fn get_register_rev(&self, register: &VarnodeData) -> Option<&[ArchRegister]> {
+        self.registers_rev.get(register).map(|list| list.as_ref())
+    }
+
+    pub fn get_register_from_offset(&self, offset: u64) -> Option<&String> {
+        self.register_offset_map.get(&offset)
+    }
+
+    pub fn get_spaces(&self) -> HashMap<SpaceName, SpaceInfo> {
+        self.sleigh.get_spaces()
+    }
+
+    pub fn endian(&self) -> ArchEndian {
+        self.sleigh.endian()
+    }
+
+    pub fn set_context_option(&mut self, option: ContextOption) {
+        // FIXME check validity (e.g. does this sla have this variable)
+        let (variable_name, variable_value) = option.value();
+        self.sleigh
+            .set_variable_default(variable_name, variable_value)
+    }
+
+    pub fn user_ops(&self) -> &[UserOpInfo] {
+        &self.user_ops
+    }
+}
+
+#[derive(Debug)]
+pub enum ContextOption {
+    ThumbMode(bool),
+}
+
+impl ContextOption {
+    fn value(&self) -> (&'static str, u32) {
+        match self {
+            ContextOption::ThumbMode(enabled) => ("TMode", *enabled as u32),
+        }
+    }
+}
+
+pub struct TranslatedPcode<'a> {
+    pcodes: &'a [Pcode],
+    bytes_consumed: u64,
+}
+
+impl<'a> TranslatedPcode<'a> {
+    pub fn pcodes(&self) -> &[Pcode] {
+        self.pcodes
+    }
+
+    pub fn bytes_consumed(&self) -> u64 {
+        self.bytes_consumed
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use styx_cpu_type::arch::arm::{variants::ArmCortexA7, ArmMetaVariants};
+    use styx_pcode_sleigh_backend::VectorLoader;
+
+    use super::*;
+
+    struct BadSla;
+    impl SlaSpec for BadSla {
+        fn spec() -> &'static [u8] {
+            &[0u8, 0u8]
+        }
+
+        fn name() -> &'static str {
+            "bad sla"
+        }
+    }
+    impl SlaRegisters for BadSla {}
+
+    /// Test when specification is invalid.
+    #[cfg_attr(miri, ignore)]
+    #[test]
+    fn test_bad_spec() {
+        let start = 0x1000;
+        let data = vec![0xFFu8; 4];
+        let load_image = VectorLoader { start, data };
+        let translator = PcodeTranslator::new::<BadSla>(
+            &ArchVariant::Arm(ArmMetaVariants::ArmCortexA7(ArmCortexA7 {})),
+            load_image,
+        );
+
+        assert!(translator.is_err())
+    }
+}
+
+#[cfg(test)]
+#[cfg(feature = "arch_arm")]
+mod arm_tests {
+    use keystone_engine::Keystone;
+    use styx_cpu_type::arch::{
+        arm::{variants::ArmCortexA7, ArmMetaVariants},
+        backends::ArchVariant,
+    };
+    use styx_pcode_sleigh_backend::{SleighTranslateError, VectorLoader};
+
+    use super::PcodeTranslator;
+    use crate::ContextOption;
+
+    fn get_asm(instr: &str) -> Vec<u8> {
+        // Assemble instructions
+        // Processor default to thumb so we use that
+        let ks = Keystone::new(keystone_engine::Arch::ARM, keystone_engine::Mode::THUMB)
+            .expect("Could not initialize Keystone engine");
+        let asm = ks
+            .asm(instr.to_owned(), 0x1000)
+            .expect("Could not assemble");
+        asm.bytes
+    }
+
+    #[cfg_attr(miri, ignore)]
+    #[test]
+    fn test_decompile_armv7() {
+        let start = 0x1000;
+        let data = get_asm("movs r2, 0xDE");
+        let load_image = VectorLoader { start, data };
+        let mut translator = PcodeTranslator::new::<styx_sla::Arm7Le>(
+            &ArchVariant::Arm(ArmMetaVariants::ArmCortexA7(ArmCortexA7 {})),
+            load_image,
+        )
+        .unwrap();
+        translator.set_context_option(ContextOption::ThumbMode(true));
+
+        let mut pcodes = Vec::new();
+        let bytes_used = translator.get_pcode(start, &mut pcodes, &mut ()).unwrap();
+
+        // arm thumb instructions are 2 bytes
+        assert_eq!(bytes_used, 2);
+        assert_eq!(pcodes.len(), 5);
+    }
+
+    #[cfg_attr(miri, ignore)]
+    #[test]
+    fn test_decompile_error() {
+        let start = 0x1000;
+        let data = vec![0xFFu8; 4];
+        let load_image = VectorLoader { start, data };
+        let mut translator = PcodeTranslator::new::<styx_sla::Arm7Le>(
+            &ArchVariant::Arm(ArmMetaVariants::ArmCortexA7(ArmCortexA7 {})),
+            load_image,
+        )
+        .unwrap();
+        translator.set_context_option(ContextOption::ThumbMode(true));
+
+        let mut pcodes = Vec::new();
+        let result = translator.get_pcode(start, &mut pcodes, &mut ());
+        assert!(matches!(result, Err(SleighTranslateError::BadDataError)));
+    }
+}
+
+#[cfg(test)]
+#[cfg(feature = "arch_ppc")]
+mod powerpc_tests {
+    use super::*;
+    use crate::sla::Ppc324xxBe;
+    use crate::VectorLoader;
+    use expect_test::{expect, expect_file};
+    use styx_cpu_type::arch::{
+        backends::ArchRegister,
+        ppc32::{Ppc32MetaVariants, Ppc32Variants},
+    };
+    use styx_pcode::pcode::VarnodeData;
+    use tap::Conv;
+
+    /// Asserts that the styx -> sla register translation does not change. This also includes spr registers.
+    #[test]
+    #[cfg_attr(miri, ignore)]
+    fn test_ppc405_register_translation() {
+        let start = 0x1000;
+        let data = vec![0xFFu8; 4];
+        let load_image = VectorLoader { start, data };
+        let mut translator =
+            PcodeTranslator::new::<Ppc324xxBe>(&Ppc32Variants::Ppc405.into(), load_image).unwrap();
+        let mut translated_registers: Vec<(ArchRegister, Option<VarnodeData>)> =
+            PcodeTranslator::get_registers_option::<Ppc324xxBe>(
+                &Ppc32Variants::Ppc405.conv::<Ppc32MetaVariants>(),
+                &mut translator.sleigh,
+            )
+            .collect();
+        // sort to ensure consistent results
+        translated_registers.sort_by_key(|(reg, _)| *reg);
+        let expected_registers = expect_file!["../expect-test-data/ppc405-registers"];
+        expected_registers.assert_debug_eq(&translated_registers);
+    }
+
+    /// Performs a snapshot test of powerpc instruction translation.
+    ///
+    /// The goal of this test is to detect changes in translation after a sla upgrade.
+    #[test]
+    #[cfg_attr(miri, ignore)]
+    fn test_ppc_translate() {
+        // objdump from example ppc program
+        // notably load/store operations are omitted because sleigh uses dynamic pointers
+        //   to represent memory spaces which change run to run.
+        let objdump = r#"
+             10c:	7c 3f 0b 78 	mr      r31,r1
+             110:	3d 20 00 00 	lis     r9,0
+             114:	39 40 00 00 	li      r10,0
+             11c:	39 20 00 00 	li      r9,0
+             124:	48 00 00 28 	b       14c <main+0x4c>
+             128:	3d 20 00 00 	lis     r9,0
+             134:	7d 4a 4a 14 	add     r10,r10,r9
+             138:	3d 20 00 00 	lis     r9,0
+             144:	39 29 00 01 	addi    r9,r9,1
+             150:	2c 09 27 0f 	cmpwi   r9,9999
+             154:	40 81 ff d4 	ble     128 <main+0x28>
+             158:	3d 20 00 00 	lis     r9,0
+             160:	7d 2f 4b 78 	mr      r15,r9
+             164:	60 00 00 00 	nop
+             168:	60 00 00 00 	nop
+             16c:	4b ff ff fc 	b       168 <main+0x68>
+             "#;
+
+        let init_pc = 0x1000u64;
+        // this jumble of iterators takes the objdump and extracts the binary from it
+        let code: Vec<u8> = styx_util::parse_objdump(objdump).unwrap();
+
+        let loader = VectorLoader {
+            start: init_pc,
+            data: code,
+        };
+        let mut translator =
+            PcodeTranslator::new::<Ppc324xxBe>(&Ppc32Variants::Ppc405.into(), loader).unwrap();
+
+        // will hold the final list of pcodes
+        let mut pcodes = Vec::new();
+        let mut pc = init_pc;
+        // translate all code to pcodes
+        while let Ok(n) = translator.get_pcode(pc, &mut pcodes, &mut ()) {
+            pc += n;
+        }
+
+        let expected_pcodes = expect![[r#"
+            [
+                IntOr Register(0x04, 4), Register(0x04, 4) -> Register(0x7C, 4),
+                IntLeft Constant(0x00, 4), Constant(0x10, 4) -> Register(0x24, 4),
+                Copy Constant(0x00, 4) -> Register(0x28, 4),
+                Copy Constant(0x00, 4) -> Register(0x24, 4),
+                Branch Ram(0x1038, 4),
+                IntLeft Constant(0x00, 4), Constant(0x10, 4) -> Register(0x24, 4),
+                IntAdd Register(0x28, 4), Register(0x24, 4) -> Register(0x28, 4),
+                IntLeft Constant(0x00, 4), Constant(0x10, 4) -> Register(0x24, 4),
+                IntAdd Register(0x24, 4), Constant(0x01, 4) -> Register(0x24, 4),
+                Copy Register(0x24, 4) -> Unique(0x14380, 4),
+                Copy Constant(0x270F, 4) -> Unique(0x14400, 4),
+                IntSLess Unique(0x14380, 4), Unique(0x14400, 4) -> Unique(0x14480, 1),
+                IntLeft Unique(0x14480, 1), Constant(0x03, 4) -> Unique(0x14500, 1),
+                IntSLess Unique(0x14400, 4), Unique(0x14380, 4) -> Unique(0x14580, 1),
+                IntLeft Unique(0x14580, 1), Constant(0x02, 4) -> Unique(0x14600, 1),
+                IntOr Unique(0x14500, 1), Unique(0x14600, 1) -> Unique(0x14680, 1),
+                IntEqual Unique(0x14380, 4), Unique(0x14400, 4) -> Unique(0x14700, 1),
+                IntLeft Unique(0x14700, 1), Constant(0x01, 4) -> Unique(0x14780, 1),
+                IntOr Unique(0x14680, 1), Unique(0x14780, 1) -> Unique(0x14800, 1),
+                IntAnd Register(0x400, 1), Constant(0x01, 1) -> Unique(0x14880, 1),
+                IntOr Unique(0x14800, 1), Unique(0x14880, 1) -> Register(0x900, 1),
+                Copy Constant(0x00, 1) -> Unique(0xB080, 1),
+                IntSub Constant(0x03, 4), Constant(0x01, 4) -> Unique(0x800, 4),
+                IntRight Register(0x900, 1), Unique(0x800, 4) -> Unique(0x900, 1),
+                IntAnd Unique(0x900, 1), Constant(0x01, 1) -> Unique(0xB080, 1),
+                BoolNegate Unique(0xB080, 1) -> Unique(0xB080, 1),
+                CBranch Ram(0xFFC, 4), Unique(0xB080, 1),
+                IntLeft Constant(0x00, 4), Constant(0x10, 4) -> Register(0x24, 4),
+                IntOr Register(0x24, 4), Register(0x24, 4) -> Register(0x3C, 4),
+                IntOr Register(0x00, 4), Constant(0x00, 4) -> Register(0x00, 4),
+                IntOr Register(0x00, 4), Constant(0x00, 4) -> Register(0x00, 4),
+                Branch Ram(0x1038, 4),
+            ]
+        "#]];
+        expected_pcodes.assert_debug_eq(&pcodes);
+    }
 }
