+++ styx/core/styx-cpu-pcode-backend/styx-pcode/src/pcode/opcodes.rs
@@ -2,7 +2,7 @@
 use num_derive::FromPrimitive;
 
 /// P-code Opcode.
-#[derive(Debug, FromPrimitive, PartialEq, Eq)]
+#[derive(Debug, FromPrimitive, PartialEq, Eq, Clone, Copy)]
 pub enum Opcode {
     ///< Copy one operand to another
     Copy = 1,
